% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parse_academic_history}
\alias{parse_academic_history}
\title{Parse the academic history output}
\usage{
parse_academic_history(df)
}
\arguments{
\item{df}{A \code{data.frame} from \code{read_academic_history}}
}
\value{
A list of checks and more subsetted data
}
\description{
Parse the academic history output
}
