% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_academic_history}
\alias{read_academic_history}
\title{Read Academic History}
\usage{
read_academic_history(file)
}
\arguments{
\item{file}{PDF file of the transcripts/application}
}
\value{
A \code{tbl} with information of the courses that were taken.
A GPA summary is given by \code{attr(output, "summary")}
}
\description{
Read Academic History
}
