% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{transform_parsed_history}
\alias{transform_parsed_history}
\title{Transform the parsed the academic history output into a \code{data.frame}}
\usage{
transform_parsed_history(out)
}
\arguments{
\item{out}{A \code{list} from \code{parse_academic_history}}
}
\value{
A single-row \code{data.frame} of the checks and passes.
}
\description{
Transform the parsed the academic history output into a \code{data.frame}
}
